+++ src/connect6.mojo
@@ -8,7 +8,7 @@ from heap import heap_add
 comptime debug = get_defined_string["ASSERT_MODE", ""]()
 
 comptime win_stones = 6
-comptime values: List[Float32] = [0, 1, 5, 25, 125, 625]
+comptime values: List[Int16] = [0, 1, 5, 25, 125, 625]
 
 
 struct Move(TMove):
