+++ src/gomoku.mojo
@@ -3,7 +3,7 @@ from traits import TGame, TMove, MoveScore
 from board import Board, Place, PlaceScore, first
 
 comptime win_stones = 5
-comptime values: List[Float32] = [0, 1, 5, 25, 125]
+comptime values: List[Int16] = [0, 1, 5, 25, 125]
 
 
 struct Move(TMove):
